getGlobalFDR

Calculate specificity as Global False Discovery Rate (FDR).

getGlobalFDR(
  seu_obj = NULL,
  features = NULL,
  tx_file = "path_to_txFile",
  cellSegMeta = "path_to_cellMeta",
  platform = NULL,
  path
)

Arguments

seu_obj

A Seurat object containing spatial data, including a path and platform attribute.

features

An optional vector of feature names (gene names) for which to calculate the FDR. If NULL (default), FDR is calculated for all features in the seu_obj.

tx_file

Path to transcription file.

platform

The platform from which the data originates. Valid options are 'Xenium', 'CosMx', and 'Merscope'. Note: 'Merscope' is currently not supported.

path

Path to generate tx file.

Value

A data frame with columns for sample_id, platform, and the calculated mean FDR across the specified features.

Details

Computes the global FDR for specified features (or all features by default) in a Seurat object. It leverages transcript localization data to calculate FDR based on the proportion of negative control or blank barcode expressions compared to the expression of actual genes.